Encapsulated JustBlackjackLogo anims into its own component
